+++ frontend/src/fetchAPI/message.ts
@@ -1,13 +1,14 @@
+import SendDM from "@src/components/main/send_dm";
 import { getToken } from "./cookie";
 import { resetCookie, getUserId } from "./cookie";
 
-export interface SendDM {
-  receivedUserId: number;
-  workspaceId: number;
+export interface SendDMForm {
+  receivedUserID: number;
+  workspaceID: number;
   text: string;
 }
 
-export interface SendDMInfo {
+export interface ResSendDM {
   id: number;
   text: string;
   sendUserId: number;
@@ -18,9 +19,8 @@ export interface SendDMInfo {
 
 const baseUrl = "http://localhost:8080/api/dm/";
 
-export async function sendDM(current: SendDM) {
+export async function sendDM(form: SendDMForm): Promise<ResSendDM|void> {
   const url = baseUrl + "send";
-  let resSendDM: SendDMInfo;
   try {
     const res = await fetch(url, {
       method: "POST",
@@ -28,14 +28,13 @@ export async function sendDM(current: SendDM) {
         Authorization: getToken(),
       },
       body: JSON.stringify({
-        receivedUserId: current.receivedUserId,
-        workspaceId: current.workspaceId,
-        text: current.text,
+        received_user_id: form.receivedUserID,
+        workspace_id: form.workspaceID,
+        text: form.text,
       }),
     });
-    resSendDM = await res.json();
+    return await res.json();
   } catch (err) {
     console.log(err);
   }
-  return;
 }
